+++ backend/index.js
@@ -77,7 +77,7 @@ io.on("connection", (socket) => {
 
 //Every minute the notification queue is updated
 cron.schedule("* * * * *", () => {
-  console.log(notificationQueue);
+  console.log("cicle");
   notificationQueue.processDueNotifications();
 });
