+++ backend/routes/appointment.js
@@ -8,6 +8,7 @@ const calculatePriority = require("../utils/calculatePriority");
 const {
   getNotificationQueue,
 } = require("../notifications/notificationQueueInstance");
+const formatDateToReadableString = require("../utils/formatDateToReadableString");
 
 //Get Patient Appointment History
 router.get("/get-all-patient-appointments/:patientId", async (req, res) => {
@@ -117,7 +118,9 @@ router.post("/create-new-appointment", async (req, res) => {
     //If the schedule date is invalid (in the past or bad format)
     //Check if the date is valid
     if (new Date(scheduleDate) < new Date()) {
-      res.status(400).json({ error: "The date for the schedule is not valid" });
+      return res
+        .status(400)
+        .json({ error: "The date for the schedule is not valid" });
     }
 
     if (existingAppointment) {
@@ -140,7 +143,9 @@ router.post("/create-new-appointment", async (req, res) => {
     /**
      * Create new notification
      */
-    const message = `You got an appointment at ${scheduleDate}`;
+    const message = `You got an appointment at ${formatDateToReadableString(
+      scheduleDate
+    )}`;
 
     //schedule a notification an hour before
     const scheduledAt = new Date(new Date(scheduleDate).getTime() - 3600000);
