+++ backend/utils/formatDateToReadableString.js
@@ -1,0 +1,15 @@
+function formatDateToReadableString(date) {
+  const options = {
+    weekday: "long",
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+  };
+
+  return date.toLocaleString(undefined, options);
+}
+
+module.exports = formatDateToReadableString;
